requirements.txt
httpx==0.28.1
beautifulsoup4==4.13.3
lxml==5.3.2
playwright==1.51.0
psycopg[binary]==3.2.6
boto3==1.38.0
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
siphash24==1.8
